Backend/requirements.txt
awscli==1.16.9
awsebcli==3.14.4
Babel==2.7.0
blessed==1.14.2
boto3==1.10.27
botocore==1.11.9
cached-property==1.5.1
cement==2.8.2
colorama==0.3.7
Django==2.2.7
django-cors-headers==3.2.0
django-jet==1.0.8
django-multiselectfield==0.1.10
django-seo-js==0.3.2
django-storages==1.8
djangorestframework==3.10.3
docker==3.7.3
docker-compose==1.21.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
jmespath==0.9.3
jsonschema==2.6.0
money==1.3.0
pathspec==0.5.5
Pillow==6.2.1
psycopg2-binary==2.8.4
pyasn1==0.4.8
python-dateutil==2.6.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.1.2
requests==2.9.1
rsa==3.4.2
s3transfer==0.1.13
semantic-version==2.5.0
six==1.12.0
sqlparse==0.3.0
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
urllib3==1.23
wcwidth==0.1.7
websocket-client==0.46.0
